requirements.txt
Django==5.0.4
Dj-rest-auth==6.0.0
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
cloudinary==1.39.1
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.2
urllib3==2.2.1
whitenoise==6.6.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.4
cryptography==42.0.5
defusedxml==0.7.1
idna==3.7
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
python3-openid==3.2.0
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1